Fix the HeapChannel.size() warning about overflowing ints prior to converting to a long (which would have required precondition violation)

+++ magpie/src/main/java/com/blackducksoftware/common/io/HeapChannel.java
@@ -134,7 +134,7 @@ public class HeapChannel implements SeekableByteChannel {
     @Override
     public long size() throws ClosedChannelException {
         requireOpen();
-        return count - off;
+        return (long) count - off;
     }
 
     @Override
